+++ libminifi/src/core/PropertyType.cpp
@@ -56,4 +56,29 @@ PropertyValue DataTransferSpeedPropertyType::parse(std::string_view input) const
   return PropertyValue::parse<core::DataTransferSpeedValue>(input, *this);
 }
 
+namespace StandardPropertyTypes {
+
+const core::PropertyType& translateCodeToPropertyType(const PropertyTypeCode& code) {
+  switch (code) {
+    case PropertyTypeCode::INTEGER:
+      return core::StandardPropertyTypes::INTEGER_TYPE;
+    case PropertyTypeCode::LONG:
+      return core::StandardPropertyTypes::LONG_TYPE;
+    case PropertyTypeCode::BOOLEAN:
+      return core::StandardPropertyTypes::BOOLEAN_TYPE;
+    case PropertyTypeCode::DATA_SIZE:
+      return core::StandardPropertyTypes::DATA_SIZE_TYPE;
+    case PropertyTypeCode::TIME_PERIOD:
+      return core::StandardPropertyTypes::TIME_PERIOD_TYPE;
+    case PropertyTypeCode::NON_BLANK:
+      return core::StandardPropertyTypes::NON_BLANK_TYPE;
+    case PropertyTypeCode::PORT:
+      return core::StandardPropertyTypes::PORT_TYPE;
+    default:
+      throw std::invalid_argument("Unknown PropertyTypeCode");
+  }
+}
+
+}  // namespace StandardPropertyTypes
+
 }  // namespace org::apache::nifi::minifi::core
